+++ src/entity/mod.rs
@@ -1,0 +1,4 @@
+pub mod extend;
+
+pub use extend::ExtendedCrud;
+pub use extend::PartialEntity;
